fix(dorion-settings): move pages to pages folder
