+++ plugins/dorion-settings/pages/PerformancePage.tsx.scss
