+++ plugins/dorion-settings/pages/PluginsPage.tsx
@@ -1,8 +1,8 @@
-import { ClientModList } from './ClientModList.jsx'
-import { PluginList } from './PluginList.jsx'
+import { ClientModList } from '../components/ClientModList.jsx'
+import { PluginList } from '../components/PluginList.jsx'
 import { css, classes } from './PluginsPage.tsx.scss'
 import { Card } from '../../../components/Card.jsx'
-import { WarningCard } from './WarningCard.jsx'
+import { WarningCard } from '../components/WarningCard.jsx'
 
 const { invoke } = (window as any).__TAURI__
