+++ plugins/dorion-settings/pages/PluginsPage.tsx.scss
